+++ src/components/RecommendedPart/RecommendedPart.styled.js
